hardening hexascii

+++ lib/hexascii/hexascii.cpp
@@ -40,12 +40,12 @@ void hexAscii::hexStringToByteArray(uint8_t *byteArrayOut, const char *hexString
     uint32_t nmbrBytes         = hexStringInLength / 2;
 
     for (uint32_t index = 0; index < nmbrBytes; index++) {
-        uint8_t leftCharacter  = hexStringIn[index * 2];
-        uint8_t rightCharacter = hexStringIn[(index * 2) + 1];
+        uint8_t leftCharacter  = static_cast<uint8_t>(hexStringIn[index * 2]);
+        uint8_t rightCharacter = static_cast<uint8_t>(hexStringIn[(index * 2) + 1]);
         leftCharacter          = toUpperCase(leftCharacter);
         rightCharacter         = toUpperCase(rightCharacter);
         if ((isHexCharacter(leftCharacter)) && (isHexCharacter(rightCharacter))) {
-            byteArrayOut[index] = (valueFromHexCharacter(leftCharacter) << 4) + valueFromHexCharacter(rightCharacter);
+            byteArrayOut[index] = (valueFromHexCharacter(leftCharacter) << 4U) + valueFromHexCharacter(rightCharacter);
         }
     }
 }
